+++ module-presentation/src/test/java/com/depromeet/image/fixture/dto/ImageUploadResponseDtoFixture.java
@@ -1,4 +1,4 @@
-package com.depromeet.memory.fixture.dto;
+package com.depromeet.image.fixture.dto;
 
 import com.depromeet.image.dto.response.ImageUploadResponse;
 import java.util.ArrayList;
