+++ module-presentation/src/test/java/com/depromeet/memory/api/PoolControllerTest.java
@@ -1,12 +1,12 @@
-package com.depromeet.memory.controller;
+package com.depromeet.memory.api;
 
 import static org.mockito.ArgumentMatchers.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-import com.depromeet.memory.config.ControllerTestConfig;
-import com.depromeet.memory.mock.WithCustomMockMember;
+import com.depromeet.config.ControllerTestConfig;
+import com.depromeet.config.mock.WithCustomMockMember;
 import com.depromeet.pool.api.PoolController;
 import com.depromeet.pool.facade.PoolFacade;
 import com.fasterxml.jackson.databind.ObjectMapper;
