+++ settings.gradle
@@ -7,3 +7,4 @@ include 'module-infrastructure:security'
 include 'module-infrastructure:object-storage'
 include 'module-infrastructure:persistence-redis'
 include 'module-infrastructure:persistence-database'
+include 'module-infrastructure:security'
